Unified type of count value passed to ST_SetKey/ST_GetKey of libgeotiff API
--HG--
extra : convert_revision : svn%3A1766ff46-f334-0410-ab20-d63176f87757/trunk%40969

+++ src/detail/reader10.cpp
@@ -263,7 +263,7 @@ bool ReaderImpl::ReadGeoreference(LASHeader& header)
         std::vector<uint8_t> data = record.GetData();
         if (uid == record.GetUserId(true).c_str() && 34735 == record.GetRecordId())
         {
-            int16_t count = data.size()/sizeof(int16_t);
+            int count = data.size()/sizeof(int16_t);
             ST_SetKey(st.get(), record.GetRecordId(), count, STT_SHORT, &(data[0]));
         }
 
@@ -275,7 +275,7 @@ bool ReaderImpl::ReadGeoreference(LASHeader& header)
 
         if (uid == record.GetUserId(true).c_str() && 34737 == record.GetRecordId())
         {
-            uint8_t count = data.size()/sizeof(uint8_t);
+            int count = data.size()/sizeof(uint8_t);
             ST_SetKey(st.get(), record.GetRecordId(), count, STT_ASCII, &(data[0]));
         }
     }
